various small fixes for 64Bit systems
texk/dvipsk: add casts to avoid compiler warnings
texk/gsftopk: add casts to avoid compiler warnings
texk/ps2pkm: import changes from libs/t1lib
texk/ttfdump: add casts to avoid compiler warnings
texk/dviljk: add casts for printing and proper format specs for scanf
             to avoid compiler warnings and potential bugs

git-svn-id: svn://tug.org/texlive/trunk/Build/source@14299 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/dvipsk/writet1.c
@@ -1593,7 +1593,7 @@ static void t1_flush_cs(boolean is_subr)
     for (ptr = tab; ptr < end_tab; ptr++) {
         if (ptr->used) {
             if (is_subr)
-                sprintf(t1_line_array, "dup %u %u", ptr - tab, ptr->cslen);
+                sprintf(t1_line_array, "dup %lu %u", (unsigned long) (ptr - tab), ptr->cslen);
             else
                 sprintf(t1_line_array, "/%s %u", ptr->name, ptr->cslen);
             p = strend(t1_line_array);
@@ -1604,7 +1604,7 @@ static void t1_flush_cs(boolean is_subr)
         else {
             /* replace unsused subr's by return_cs */
             if (is_subr) {
-                sprintf(t1_line_array, "dup %u %u%s ", ptr - tab, cs_len,
+                sprintf(t1_line_array, "dup %lu %u%s ", (unsigned long) (ptr - tab), cs_len,
                         cs_token_pair[0]);
                 p = strend(t1_line_array);
                 memcpy(p, return_cs, cs_len);
